SIMPLY-3608 Remove bookmarks code duplication, minor cleanup
